(events) Update 12 Days of Chocolatey Page
The 12 Days of Chocolatey is now over, so any details about it being an
upcoming live event have been removed. The YouTube replay links have
been replaced by links that lead to the video in our resource area.
Also, thumbnail images of each video are now featured in the schedule
area, instead of the date calendar.