(GH-1039) Upgrade project to .NET Framework 4.5.2
In order to accomplish this, it was necessary to upgrade EntityFramework
package to 5.0.0.  This was due to the fact that some attributes that
are being used on the entity classes have moved to a new namespace.
The necessary attribute changes have been made.

In addition, it was necessary to add the NotMapped attribute to a number
of properties, otherwise EF was changing that a migration was required
to be added.  My assumption here is that in EF 5.0.0 there has been a
change in how mapped properties were discovered. In EF 4.3.1 the
NotMapped attribute was not required, but now it is.

Where necessary, NuGet packages were uninstalled and re-installed in
order to re-target to .NET Framework 4.5.2.